noxfile.py: Rerun 'poetry export' if reqs_from_poetry.txt is missing
I've run into a few cases where I've needed to recreate the Nox
virtualenv, and I've simply deleted the corresponding
reqs_from_poetry.txt file to ensure that everything is regenerated from
scratch. However, the test in noxfile.py only checks for the associated
.hash digest file, and then promptly fails when it tries to reuse the
missing file.

Add another clause to the if-condition to fix this.