Add 'packaging' as a core dependency
We're about to replace our dependency on pkg_resources (from setuptools)
with packaging.requirements instead.

Coincidentally, packaging is already a transitive dependency of ours
(via pip-requirements-parser), so this does not add any additional bloat
when installing FawltyDeps.

We require >=24.0 of this package even though there already exists a
later 24.1 release. The reason for this is that 24.0 is the last release
to support Python v3.7.