parse_pyproject_toml: Support Pixi dependencies
Add preliminary support for parsing Conda dependencies in the
[tool.pixi.dependencies] section of pyproject.toml. This is used by Pixi
to declare mandatory dependencies that are taken from the Conda
ecosystem (rather than the PyPI ecosystem).

One complication of Pixi dependencies is that if a dependency is listed
_both_ as a PyPI dependency (in the standard PEP621 location) and as a
Conda dependency in the [tool.pixi.dependencies] table, then the latter
takes precedence.

NOTE: We do not currently differentiate between Conda dependencies and
PyPI dependencies, meaning that we assume that a Conda dependency named
FOO will map one-to-one to a Python package named FOO. This is certainly
not true for Conda dependencies that are not Python packages, and it
probably isn't even true for all Conda dependencies that do indeed
include Python packages.

For now, we only make sure that the "same" dependency is not yielded
twice when it appears in both as a PyPI and Conda dependency.