parse_pyproject_toml: Ignore Pixi's self-dependency
Pixi inserts this self-dependency by default in its projects:

    ```toml
    [project]
    name = "my_project"
    requires-python = ">= 3.9"

    [tool.pixi.project]
    channels = ["conda-forge"]
    platforms = ["linux-64"]

    [tool.pixi.pypi-dependencies]
    my_project = { path = ".", editable = true }
    ```

This self-dependency should NOT be considered a declared dependency,
but is rather a mechanism to produce an editable installation of the
project itself inside the environments that Pixi creates.

In general it probably does not make sense for FawltyDeps to accept
these self-dependencies, no matter how/where they are found?