extract_declared_dependencies.py: Add support for parsing pixi.toml
This adds support for the default configuration file used in Pixi
project: pixi.toml. The relevant fields in this file mirror the Pixi
fields we already parse in pyproject.toml (but without the tool.pixi
prefix). Thus we can reuse the parsing functions we already wrote for
supporting Pixi projects via pyproject.toml.