sample_projects: Add Pixi project w/default settings
This was created with the following commands:

  - pixi init pixi_default_example
  - cd pixi_default_example/
  - pixi add python
  - pixi add --pypi requests
  - pixi add ninja
  - echo "import requests" > main.py

followed by removing/truncating files that I do not currently consider
to be important for FawltyDeps to make sense of this project.

The result is a simple Pixi project with two declared Conda
dependencies (Python itself + ninja) and one PyPI dependency
(requests), along with a single source file that imports requests.

The expected result of running FawltyDeps here is reporting ninja as an
unused dependency. This is because ninja is a Conda package with no
Python components, and since it is not found as a Python package, we
fall back to the identity mapping and expect it to be imported as
"ninja".

We should fix this by digging into the Conda environment to discover
that ninja has no Python components, and therefore is not expected to be
imported.