sample_projects: Add Pixi project using pyproject.toml
This was created with the following commands:

  - pixi init --format pyproject pixi_pyproject_example
  - cd pixi_pyproject_example/
  - pixi add python
  - pixi add requests
  - pixi add --pypi fawltydeps
  - echo "import requests" > pixi_pyproject_example/__init__.py

followed by removing/truncating files that I do not currently consider
to be important for FawltyDeps to make sense of this project.

The result is a simple Pixi project with two declared Conda
dependencies (Python itself + requests) and one PyPI dependency
(fawltydeps), along with a single source file that imports requests.

The expected result of running FawltyDeps here is success (requests is
both declared and imported, fawltydeps is part of the default tool
exception, and Python itself is auto-ignored as a dependency).