extract_declared_dependencies: Handle exceptions due to invalid toml
When parsing a pyproject.toml or pixi.toml with invalid TOML data, the
tomllib parser will raise an exception that is not properly caught or
handled by FawltyDeps.

Fix this by logging an error message and returning from the parse
function. This allows FawltyDeps to continue parsing other files and
reporting undeclared/unused dependencies, whereas before it would abort
with a traceback.