PyEnvSource: Slightly loosen restrictions on Python environments
For a Python environment to be considered valid, we typically look for a
Python executable, and an associated site-packages directory. The check
for the Python executable uses .is_file(), which returns true as long
the Python executable is a regular file, or a symlink to an existing
regular file.

We are however running into occasional CI failures where the Python
executables in some of our (cached) virtualenvs for real_projects appear
to be broken symlinks.

I suspect that environments where Python executables are broken symlinks
might also occur in other circumstances (e.g. on Nix where the
underlying Python executable might have been garbage-collected due to a
missing GC root, or on a more regular Linux distro where a stale
virtualenv might be pointing to a system Python that has since been
upgraded).

In any case, a Python environment with a broken python symlink might be
non-functional, but should that really cause problems for FawltyDeps?
Since FawltyDeps can traverse the environment without relying on the
environment's Python executable, the answer is actually no: All we
really need is for bin/python to serve as a marker that this is (or at
some point was) a Python environment, and a broken symlink should not
stop us.

Hence tweak our code to take broken symlinks into account when
recognizing Python environments.