packages: Provide suggest_packages() as a public function
Instead of hiding _candidates() inside the check module, make it
available as a public function in the packages module. This new
function somewhat mirrors the existing resolve_dependencies() function,
and we might even consider collecting this (and setup_resolvers()) into
a future ResolverStack class, to abstract access to operations that
involve multiple configured resolvers.