UndeclaredDependency: Convert .candidates from list[Package] to set[str]
We don't need/want to reference entire Package objects from
UndeclaredDependency objects. It complicates the hierarchy of class
composition by making a more complex class (Package) part of a simpler
class (UndeclaredDependency). Furthermore, at the point where we have
determined the .candidates for an UndeclaredDependency, we are merely
interested in reporting the package names, and we no longer need access
to the other details in a Package object.

This allows us to avoid reorganization of fundamental classes in order
to make type checking work without circular imports.