Accept *-wildcards for ignore_undeclared and ignore_unused
The '*' wildcards are translated into `.*` regular expression tokens
(which match anything), and then passed to re.fullmatch(), which I believe
should do exactly what we want in this case. That said, we need to take
extra care to avoid injection of regexp special chars in the ignore
patterns, so all other parts of the pattern are filtered through
re.escape().