Suggest original (non-normalized) package names
Until now, we've only used normalized package names internally, as this
is how they are resolved and used by most parts of the Python packaging
infrastructure. However, when we suggest package names for the user to
declare, we should rather use the original package name as it is
presented in the package's own metadata.

In other words, we want to suggest installing "Scikit-Learn" or "SciPy"
(if that is how those packages choose to present itself) instead of
"scikit_learn" or "scipy".

Thus, we now store the original package name as given to the Package
constructor in the .package_name member. This is the name we will then
present when this package is being suggested as a potential declaration
for an undeclared dependency.

We add a .normalized_name cached property for computing the normalized
name, but this is in fact never used (outside of tests): Package objects
are typically kept in dictionaries that are keyed by their normalized
name, so objects are typically looked from their normalized name, rather
than the normalized name being needed once the object is found.