Adjust whitespace/indentation in human-readable output
Before, with --detailed:

  These imports appear to be undeclared dependencies:
  - 'other_import' imported at:
      path/to/file.py
      path/to/other_file.py
  - 'some_import' imported at:
      path/to/file.py
      path/to/other_file.py
  - 'yet_another' imported at:
      path/to/file.py
      path/to/other_file.py

After, with --detailed:

  These imports appear to be undeclared dependencies:
  - 'other_import'
      imported at:
        path/to/file.py
        path/to/other_file.py
  - 'some_import'
      imported at:
        path/to/file.py
        path/to/other_file.py
  - 'yet_another'
      imported at:
        path/to/file.py
        path/to/other_file.py

(and similar for unused dependencies). No change in output when
--detailed is not used.

The objective here is to make room for different kinds of detail to be
added to the output, for now the only plan is to add package name
suggestions for undeclared dependencies (see next commit).