test_package: Clean up test parameters
These tests included ["and", "other", "names"] and similar in the test
parameters to represent the fact that .is_used() happens to be called
with the superset of all import name found in the code, and that it is
supposed to find matching import names (and return `true`) even when
most of the given import names do not match.

Instead of polluting every single test case with these extra names, add
a single test case at the end to verify this part of the expected
behavior, and let the remainder of the tests focus on the exact behavior
that they're meant to test.