cli_parser.py: Rename basepaths to search_paths
We accept only one kind of positional arguments, named 'basepaths'
internally, that serve as a default for --code/--deps/--pyenv.

With the introduction of a separate --base-dir option for directly
controlling the base directory used when determining 1st- vs 3rd-party
imports, it becomes clear that 'basepaths' is no longer an appropriate
name for the positional arguments:

The relationship between 'basepaths' and the new --base-dir option is
indirect, at best: 'basepaths' sets the default paths (dirs or files)
that will be searched/traversed (modulo overrides from --code, --deps,
or --pyenv), and the base directory for 1st-party imports defaults to
any directories passed into the traverse machinery.

Thus, when file paths are passed via 'basepaths', there is actually no
base directory information (and we fall back to the current directory),
and when directories are passed to 'basepaths', they are still subject
to overrides via --code.

Therefore, rename this internal concept to 'search_paths' as that more
closely illustrates what it does: It determines the paths (dirs or
files) that will be searched for code/deps/pyenvs (when not overridden
by the respective options).