traverse_project: Apply --base-dir to directory traversals as well
The --base-dir option was originally conceived as a way to communicate
base directory information (i.e. where to find 1st-party imports) when
only direct file paths were given to --code.

However, when a user specifies --base-dir, it makes sense that this is
_also_ obeyed when directory traversal actually _does_ take place. Let's
use the bundled sample project test as an example:

Imagine a (mono)repo with these code files:
  - subproject/lib/foo.py
    - a 1st-party/internal module that declares `class Bar`
  - subproject/main/main.py
    - the main subproject that we want to analyze,
      it uses the above module via: `from lib.foo import Bar`
  - subproject/other/other.py
    - unrelated subproject that is not part of this analysis

One way to perform this analysis with FawltyDeps, is to pass the exact
file paths we want to analyze:

    fawltydeps --code subproject/main/main.py subproject/lib/foo.py

The `--base-dir` option was specifically added for this scenario, as
with direct file paths, there is no way to deduce the correct base
directory that FawltyDeps needs in order to resolve 'lib' as a 1st-
party import. With `--base-dir subproject/` this is resolved.

However, there is another variant of this scenario:

    fawltydeps --code subproject/main subproject/lib

Here, we do have some directory context, but it is wrong: the 'lib'
import in `main.py` will be resolved using `subproject/main` as the
implicit base directory, where it will not be found.

Again, the solution should be the same from the user's perspective:
Pass `--base-dir subproject/` to tell FawltyDeps where to find 1st-party
imports.

With this commit, we allow settings.base_dir to override the implicit
base directory passed via --code, so that we can behave as expected in
this second case.