Downgrade the use of code+deps+pyenv+search_paths to a warning
The search_paths positional argument is not a proper setting on its own,
but rather serves as a default for --code, --deps, and --pyenv. When all
those are specified, search_paths has no effect. This was checked in the
code, and an error was raised, telling the user not to use all four.

However, with more thought, a warning is more appropriate here: The
documentation is clear that --code/--deps/--pyenv take precedence over
search_paths, so there is no question about how FawltyDeps should behave
in this scenario, rather it's more a question of helping to clear up any
confusion on the user's part. Hence a warning is more appropriate than
aborting with an error message.

Adjust associated test case accordingly.