extract_imports.parse_source(): Fix case when base_dir is not a parent
With the new --base-dir CLI argument, we now have a situation where
parse_source() can be passed a 'src' CodeSource object with a custom
src.base_dir that is _not_ a parent (or grandparent) directory of
src.path.

We used to pass the directories between src.path and src.base_dir
(calculated by dirs_between()) as additional source paths into isort,
to help with correctly placing 1st-party imports. When the
src.base_dir is not a parent, we can no longer do this.

While fixing this code, also update dirs_between() to raise a ValueError
when the passed arguments are not as expected (parent being a parent
directory of chile), and rewrite it from recursive to iterative style.