extract_imports.parse_source: Improve interaction with dirs_between()
There is no need for dirs_between() to raise an exception, when all
we're looking to do in the exceptional case is to replace the returned
paths with nothing (i.e. an empty tuple).

Instead, rephrase the implementation of dirs_between() to a simpler
one-liner that naturally does the Right Thing (tm) when the two given
paths are not directly related. Hence we can avoid the try-except clause
in the parse_source() caller.

Note: The only other user of dirs_between() always ensures that the two
given paths are directly related, so this refactoring does not affect
that use case.

Co-authored-by: Johan Herland <johan.herland@tweag.io>